Moving disruption controller e2es to workload/
